requirements.txt
streamlit>=1.38.0
requests>=2.31.0
deep-translator>=1.11.4
pandas>=2.1.0
bcrypt>=4.1.2
python-dotenv>=1.0.1
feedparser>=6.0.11
